requirements.txt
requests==2.31.0
pandas==2.1.4
pandas-ta==0.3.14b0
numpy==1.26.3
matplotlib==3.8.2
tabulate==0.9.0
python-dotenv==1.0.0
